增加 排序功能!

+++ src/com/sht/entity/articategory/ArtiCategoryEntity.java
@@ -35,7 +35,7 @@ public class ArtiCategoryEntity implements java.io.Serializable {
 	private java.lang.String categoryName;
 
 	/**排序*/
-	private java.lang.Integer orders;
+	private double orders;
 	
 	/**
 	 *方法: 取得java.lang.String
@@ -75,11 +75,11 @@ public class ArtiCategoryEntity implements java.io.Serializable {
 	}
 
 	@Column(name ="ORDERS",nullable=true,length=11)
-	public Integer getOrders() {
+	public double getOrders() {
 		return orders;
 	}
 
-	public void setOrders(Integer orders) {
+	public void setOrders(double orders) {
 		this.orders = orders;
 	}
 }
