+++ src/com/sht/entity/selectedarticles/SelectedArticlesEntity.java
@@ -38,7 +38,7 @@ public class SelectedArticlesEntity implements java.io.Serializable {
 	/**下架时间*/
 	private java.util.Date closingTime;
 	/**排序*/
-	private java.lang.Integer orders;
+	private double orders;
 	
 	/**
 	 *方法: 取得java.lang.String
@@ -97,7 +97,7 @@ public class SelectedArticlesEntity implements java.io.Serializable {
 	 *@return: java.lang.String  排序
 	 */
 	@Column(name ="ORDERS",nullable=true,length=11)
-	public java.lang.Integer getOrders(){
+	public double getOrders(){
 		return this.orders;
 	}
 
@@ -105,7 +105,7 @@ public class SelectedArticlesEntity implements java.io.Serializable {
 	 *方法: 设置java.lang.String
 	 *@param: java.lang.String  排序
 	 */
-	public void setOrders(java.lang.Integer orders){
+	public void setOrders(double orders){
 		this.orders = orders;
 	}
